Move endpoint to be under job
