+++ command/agent/job_endpoint_test.go
@@ -443,3 +443,42 @@ func TestHTTP_JobAllocations(t *testing.T) {
 		}
 	})
 }
+
+func TestHTTP_PeriodicForce(t *testing.T) {
+	httpTest(t, nil, func(s *TestServer) {
+		// Create and register a periodic job.
+		job := mock.PeriodicJob()
+		args := structs.JobRegisterRequest{
+			Job:          job,
+			WriteRequest: structs.WriteRequest{Region: "global"},
+		}
+		var resp structs.JobRegisterResponse
+		if err := s.Agent.RPC("Job.Register", &args, &resp); err != nil {
+			t.Fatalf("err: %v", err)
+		}
+
+		// Make the HTTP request
+		req, err := http.NewRequest("POST", "/v1/job/"+job.ID+"periodic/force", nil)
+		if err != nil {
+			t.Fatalf("err: %v", err)
+		}
+		respW := httptest.NewRecorder()
+
+		// Make the request
+		obj, err := s.Server.JobSpecificRequest(respW, req)
+		if err != nil {
+			t.Fatalf("err: %v", err)
+		}
+
+		// Check for the index
+		if respW.HeaderMap.Get("X-Nomad-Index") == "" {
+			t.Fatalf("missing index")
+		}
+
+		// Check the response
+		r := obj.(structs.PeriodicForceResponse)
+		if r.EvalID == "" {
+			t.Fatalf("bad: %#v", r)
+		}
+	})
+}
