--- command/agent/periodic_endpoint.go
@@ -1,40 +1,0 @@
-package agent
-
-import (
-	"net/http"
-	"strings"
-
-	"github.com/hashicorp/nomad/nomad/structs"
-)
-
-func (s *HTTPServer) PeriodicSpecificReqeust(resp http.ResponseWriter, req *http.Request) (interface{}, error) {
-	path := strings.TrimPrefix(req.URL.Path, "/v1/periodic/")
-	switch {
-	case strings.HasSuffix(path, "/force"):
-		jobName := strings.TrimSuffix(path, "/force")
-		return s.periodicForceRequest(resp, req, jobName)
-	default:
-		return nil, CodedError(405, ErrInvalidMethod)
-
-	}
-
-}
-
-func (s *HTTPServer) periodicForceRequest(resp http.ResponseWriter, req *http.Request,
-	jobName string) (interface{}, error) {
-	if req.Method != "PUT" && req.Method != "POST" {
-		return nil, CodedError(405, ErrInvalidMethod)
-	}
-
-	args := structs.PeriodicForceRequest{
-		JobID: jobName,
-	}
-	s.parseRegion(req, &args.Region)
-
-	var out structs.PeriodicForceResponse
-	if err := s.agent.RPC("Periodic.Force", &args, &out); err != nil {
-		return nil, err
-	}
-	setIndex(resp, out.Index)
-	return out, nil
-}
