+++ website/source/layouts/http.erb
@@ -97,10 +97,6 @@
 					<a href="/docs/http/status.html">Status</a>
                 </li>
 
-				<li<%= sidebar_current("docs-http-periodic") %>>
-					<a href="/docs/http/periodic.html">Periodic Jobs</a>
-                </li>
-
 			</ul>
 		</div>
 	<% end %>
